Add custom Donut Chart with total quantity
This component is heavily based on the Shadcn UI pie charts.
Shadcn UI charts are based in Recharts.
This chart is a donut chart with a text in the middle, representing the total of all quantities.
It is available in 4 different sizes: "sm", "md", "lg" and "xl".
Additionally you can set custom size properties.
It can be used in a financial related context (adding the currency symbol to the total) or everywhere else (no currency symbol).
It has a Title and a Subtitle. As well as a subtitle for the total value.
At the bottom you can add a description consisting of a title and subtitle.
The data and the data format can be passed down to the component from the parent component.