Add raw material API endpoint
The API endpoint will be used for fetching and returning all information related to raw materials.

For the beginning the focus will be on gold price(s).
The prices can be queried for the euro and usd denomination.
The prices endpoint will return the aggregated gold prices for a given timeRange.
This is implemented by truncation the timestamp of each price data point to the specific format relevant to the given time range.
Afterwards the closing price of each timestamp set is taken as representative of the whole set.
For example:
For the time range of a week, we want to have the closing price of each day.
Therefore each timestamp gets reduced to the specific day by a SQL query and then the last entry for each day is taken.
The db data timestamps are standard time. Therefore the local time is converted into standard time for the query. Afterwards the results will be transformed back into local time.