Feature | Implement metals price query for a given time range and a given aggregationInterval
Main points:
1) The function naming was changed ti be consistent with the API endpoint schema
2) The API endpoint schema was simplified
We only leave the metal parameter, the other variables become query params.
3) The API endpoint "getPricesRange" is quite similar to the "getPricesHistorical" query
The differences are that the timeRange is now determined by a start and end date and the aggregationInterval can be set freely.
The startDate and the endDate need to be in standard time. This is explicitly stated in the Zod description.

Quick Fixes:
Change "PeriodDuration" type to both new types "TimeRange" and "TimeInterval"