[Fix] Correct financeService functions
A minor mistake of using - instead of / for the calculation of the gain duration with compound interest has been fixed.

The issues because of time differences have been resolved.
Now you get the correct amount of milliseconds passed for "YTD" and "ALL".
To ensure proper conversion to a UTC Date, Date.UTC has been used.

The MS_PER_xxx have been given exact values instead of being calculated.
These values are the exact same that Postgres uses for their interval attributes.
This is essential for us, as we are using a Postgres DB and we might want to switch back and forth between computing on the server vs. locally while still ensuring consistency.