[Feature] Implement custom Sign-In and Sign-Up pages
Instead ob being taken to an external site, now the user stays on the trusted website.

The concept of public routes has been introduced.
In general this is a webapp supposed to only e accessible to users with an account.
So by default the sites should only be generated, if the user is signed in, otherwise nothing will be shown and the user is redirected to the Sign-In page.
Exceptions are the Sign-In and Sign-Up pages themselves and the homepage, as this is supposed to be used for a landing page to sign up to the webapp.

As protection in the middleware is not safe in general, the route protection is only meant as an additional safety layer.
All pages that are not public should include Clerks <Protect> component for guaranteed safety.
Therefore the page is divided into to different transparent route folders.
(Website) is for public pages that are there to introduce and explain the product to potential users and maybe some free features for advertising.
(Root) is for everything else - all non public pages.
These come in addition to the (Api) transparent route.
Now every private page automatically uses the (Root) layout which already includes Clerks <Protect> component.
Therefore is must not be added manually every time, as this would pose the risk of it being forgotten.