[Format] Adjust and refactor sidebar an header to new (Website) and (Root) split
The navbar has been adjusted to only be shown if the user is signed in.

The header has been adjusted to show only the Sign-In and Sign-Out button if the user is not signed in.
If he is signed in, the header shows the toggle for the sidebar and will display the breadcrumbs path of the site the user is currently on.

The navigationSidebar has been wrapped to fetch the user data like name and email before rendering the sidebar itself.
This user information is used and displayed in the sidebar user menu.

The navigationProjects and navigationPortfolio components have been renamed to be suitable for a more generic purpose.
The now describes the functionality of that component instead of the use case it is used for.